Heater stability fixes
The code now uses SendOnlySoftwareSerial. This can easily be changed to use the built in SoftwareSerial if you want to.
Other changes are largely around stability of the system in regards to when it automatically resets to try and keep the heater operational without any user intervention. Change the timing values as best suit your needs.